add a boundary rectangle argument for "ftk_canvas_draw_string"
* it fixes the following bug:

when one of clip regions contains and is larger than the boundary
rectangle of the string, it's possible some characters of the string
have been out of the boundary rectangle but they still be drawed.

and when the clip region is the widget itself, the characters which
are out-of-box will not be drawed.

so without the fix, the behavior of "ftk_canvas_draw_string" is inconsistent.
demo_label.c is a example. In FTK_MAIN if using "ftk_widget_set_text" to
set the text of the info label, the text will be drawed, but in
"timeout", the text will Not be drawed.

* it also introduce the following enhancement:

it check the intersection between the current clip region and the string
boundary rectangle first, and if there is no intersection, it just
return. so it gets rid of the two-level for loop when there is no
intersection.